+++ src/core/cache/sharedCaches.ts
@@ -23,5 +23,5 @@ export const globalWordCache = new Cache<string, GlyphData>();
 
 export const globalClusteringCache = new Cache<
   string,
-  { glyphIds: number[]; groups: number[][] }
+  { glyphIds: number[]; positions: { x: number; y: number }[]; groups: number[][] }
 >();